% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateFooter}
\alias{CreateFooter}
\title{Report footer}
\usage{
CreateFooter()
}
\description{
Report footer
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
