% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_general.R
\name{GetRCommandHistory}
\alias{GetRCommandHistory}
\title{Export R Command History}
\usage{
GetRCommandHistory()
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Export R Command History
}
