% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformMetaEffectSize}
\alias{PerformMetaEffectSize}
\title{Perform combine effect size meta-analysis method}
\usage{
PerformMetaEffectSize(method = "rem", BHth = 0.05)
}
\arguments{
\item{method}{Method used, either "rem" or "fem"}

\item{BHth}{P-value threshold}
}
\description{
Perform combine effect size meta-analysis method
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
