% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformMetaMerge}
\alias{PerformMetaMerge}
\title{Perform direct merging meta-analysis}
\usage{
PerformMetaMerge(BHth = 0.05)
}
\arguments{
\item{BHth}{P-value threshold for the significance level}

\item{method}{Method used, either "rem" or "fem"}
}
\description{
Directly merge all datasets and analyze as single data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
