% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{PerformNormalization}
\alias{PerformNormalization}
\title{Perform data normalization}
\usage{
PerformNormalization(
  dataName,
  norm.opt,
  var.thresh,
  count.thresh,
  filterUnmapped
)
}
\arguments{
\item{norm.opt}{Normalization method to be used}

\item{var.thresh}{Variance threshold}

\item{count.thresh}{Count threshold for RNA-seq data and abundance threshold for microarray data}

\item{filterUnmapped, }{boolean, whether to filter unmapped genes or not}

\item{abundance}{Relative abundance threshold}
}
\description{
Filtering and Normalizing gene expression data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
