% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformPvalCombination}
\alias{PerformPvalCombination}
\title{Perform combine p-value meta-analysis}
\usage{
PerformPvalCombination(method = "stouffer", BHth = 0.05)
}
\arguments{
\item{method}{Method used, either "stouffer" or "fisher"}

\item{BHth}{P-value threshold}
}
\description{
Perform combine p-value meta-analysis
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
