% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformVoteCounting}
\alias{PerformVoteCounting}
\title{Perform vote counting meta-analysis method}
\usage{
PerformVoteCounting(BHth = 0.05, minVote)
}
\arguments{
\item{BHth}{P-value threshold for the significance level}

\item{method}{Method used, either "rem" or "fem"}
}
\description{
Perform vote counting meta-analysis method
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
