% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_express_meta.R
\name{SanityCheckData}
\alias{SanityCheckData}
\title{Sanity check individual dataset for meta-analysis}
\usage{
SanityCheckData(fileName)
}
\arguments{
\item{fileName}{The filename of dataset in qs format}
}
\description{
Sanity check individual dataset for meta-analysis
}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
McGill University, Canada
License: MIT
}
