% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_express_meta.R
\name{SanityCheckMeta}
\alias{SanityCheckMeta}
\title{Sanity check metadata after metadata edited}
\usage{
SanityCheckMeta(fileName, init)
}
\description{
Sanity check metadata after metadata edited
}
